requirements.txt
pyserial>=3.5
pymodbus>=3.0.0
matplotlib>=3.5.0
